Process type names
